only execute JobExecution initializer in non-test environments

+++ config/initializers/job_execution.rb
@@ -1,22 +1,22 @@
 JobExecution.setup
 
-Rails.application.config.after_initialize do
+unless Rails.env.test?
   Job.pending.each do |job|
     JobExecution.start_job(job.deploy.reference, job)
   end
-end
 
-Signal.trap('USR1') do
-  # Disable new job execution
-  JobExecution.enabled = false
-  sleep(5) until JobExecution.all.empty?
+  Signal.trap('USR1') do
+    # Disable new job execution
+    JobExecution.enabled = false
+    sleep(5) until JobExecution.all.empty?
 
-  # Pass USR2 to the underlying server
-  Process.kill('USR2', $$)
-end
+    # Pass USR2 to the underlying server
+    Process.kill('USR2', $$)
+  end
 
-at_exit do
-  Deploy.active.each do |deploy|
-    deploy.stop!
+  at_exit do
+    Deploy.active.each do |deploy|
+      deploy.stop!
+    end
   end
 end
